Fix type and lint error
